moved all information related commands to their own folder
