+++ commands/info/checkcards.js
@@ -1,6 +1,5 @@
-const { models } = require('../database');
-const { findplayer} = require('./findplayer');
-const {parseCartas} = require('../helpers/');
+const { findplayer} = require('../findplayer');
+const helpers = require('../../helpers');
 
 async function checkcards(player, message, args) {
     const mentionedMember = message.mentions.members.first();
@@ -33,7 +32,7 @@ async function checkcards(player, message, args) {
 
     }    
     result.push(`Estas son ${selfcheck ? `tus cartas` : `las cartas de <@${member.id_discord}>`} similares a \"${name}\"`);
-    result.push(parseCartas(cartas, showRepeats = true));
+    result.push(helpers.parseCartas(cartas, showRepeats = true));
     return result;
 }
