+++ commands/info/index.js
@@ -1,0 +1,11 @@
+const { album } = require('./album');
+const { checkcards } = require('./checkcards');
+const { checkseries } = require('./checkseries');
+const { info } = require('./info');
+
+module.exports = {
+    album,
+    checkcards,
+    checkseries,
+    info,
+}
